mwm: Always set the border width
The border width is not set in _mwm_configure_request() if the client
did not request a specific border width. That means that clients that
leave their border width unspecified will end up with the default
border width (which is not zero).
This commit modifies _mwm_configure_request() to always set the
border width of a client to 0.