monitor: Don't XMapRaise() statusbar every time it is redrawn
The statusbar is mapped raised everytime it is redrawn, even if it is
obscured by another window. If the other window is not tiled (that is,
its position and size are not controlled by mwm), it will try to bring
itself back to the front, and the two windows end up competing to be
on top of the stack.
This commit makes sure the statusbar is mapped only once, when it is
initialized. Since statusbars are never unmapped, there is no need to
map them every time they are redrawn.