monitor: Update indicator geometry when monitor geometry changes
The indicator geometry is not updated when the monitor geometry is
changed, causing the indicator size and/or position to be wrong when
the resolution or position of a monitor is changed.
This commit makes sure the indicator geometry is updated when the
geometry of the containing monitor is changed.