Don't install/uninstall docs from src Makefile
The documentation is installed and uninstalled by the install and
uninstall rules of the Makefile in the src directory, even though
the documentation is not in that directory.

This commit adds a Makefile to the docs repository so that the src
Makefile does not have to install documentation which is not part
of the src directory.