Handle PropertyNotify events on clients
Because mwm does not handle PropertyNotify events on windows other
than the root window, transient clients are not redrawn as needed.

This commit modifies mwm so that it handles PropertyNotify events
on client windows, redrawing transient windows and handling window
manager hints. Size hints on clients are ignored, since we do not
allow clients to be resized.