Remember the pointer position when shifting focus
When focusing on a client, mwm moves the pointer to the center of
that client. However, when performing tasks that require switching
back and forth between clients, it would be more convenient if the
pointer was moved back to its original position.

This commit modifies mwm so that it saves the pointer position for
each client. When shifting the focus, the pointer position is saved
in the current client, and the position saved in the new client is
restored after the shift. If the client was resized after it lost
the focus, the pointer position will be scaled appropriately.