xrandr: Report correct width and height of rotated crtcs
When a crtc is rotated 90 or 270 degrees, the width and height in
structures describing the geometry of the crtc is not flipped as
expected by mwm. This causes clients to be arranged incorrectly on
rotated crtcs.

This commit modifies mwm so that it flips the width and height of
a monitor if the underlying crtc is rotated 90 or 270 degrees.