mwm: Add method for looking up non-standard atoms
To lookup atoms that are not defined in the X11 specification,
the `mwm_get_atom_by_name()` method has to be used, which queries
the X Server for every call.

This commit adds the `mwm_get_atom()` method, which can be used to
look up atoms from a set of non-standard atoms that are cached at
startup.