mwm: Handle arbitrary-length text properties
The `mwm_get_status()` and `mwm_get_text_property()` methods require the
caller to allocate a buffer, making them incapable of handling text
properties of arbitrary sizes.

This commit modifies the `mwm_get_text_property()` and `mwm_get_status()`
methods to that they write data to a heap-allocated buffer that is passed
to the caller.