mwm: Avoid rendering text outside of windows
When the `mwm_render_text()` and `mwm_render_text_vertical()` are
passed a string that, when rendered, is longer than the window that
it is rendered on, it will be rendered beyond the borders of the
window.

This commit adds arguments for the maximum width and height to the
`mwm_render_text()` and `mwm_render_text_vertical()` functions,
allowing the caller to specify the dimensions of the window that
the text is rendered on.