mwm: Fix calculation of invalid/default command name
When mwm is compiled with debug output, it will print the names of
commands that are executed. However, the calculation for the index
of the default entry (for invalid command ids) is incorrect,
resulting in an out-of-bounds read.

This commit fixes the calculation of the default index so that the
last entry of the command name array is printed if an invalid
command id was executed.