mwm: Change focused monitor only when redrawing
When the user requests the focus to be moved to another monitor,
the focus is changed immediately, causing other operations to affect
the newly focused monitor, instead of the one that was focused before.

This commit modifies mwm so that it postpones changing the monitor
focus until the time when it redraws the monitors. This prevents
operations from affecting the wrong monitor.