layout: Add sink layout
The available layouts divide the workspace either horizontally or
vertically, which is not ideal when viewing more than two clients
that have different content orientation.

This commit adds the `sink` layout which arranges the clients in a
swirl similar to depictions of the golden ratio, or things going
down the drain (hence the name).