workspace: Double-buffer the focused client
Workspaces have a pointer to the focused client that is directly
updated when the focus is changed, causing subsequent events to
affect the new focus, even though the focus change is not visible
to the user yet.

This commit modifies the workspace implementation so that it uses
a double-buffering approach: it keeps a pointer to the client that
is currently focused, and the client that will receive the focus
next. The new `workspace_update_focus()` method is used to make a
focus change effective, and should be called before redrawing a
workspace.