Use a global mwm instance
The mwm type is implemented in a way that it can be instantiated
multiple times, even though this is highly unlikely to ever happen.
This makes it necessary to carry and pass around pointers to mwm
instances, even though all of them point to the same instance.

This commit modifies mwm so that it uses only one, globally shared
mwm instance.