client: Make `client_redraw()` resize the client
The `client_redraw()` method changes the client position, but it does
not resize the client, thus requiring a call to `client_show()` to
make all client changes visible.

This commit modifies the `client_redraw()` method so that it uses the
`XMoveResizeWindow()` function instead of `XMoveWindow()`, resizing
and moving the client in the same call.