mwm: Clean up double-buffering of monitor focus
The way mwm handles double-buffering of the focused monitor differs
from the approach used for client geometries. To make the code more
readable, the latter approach should be used.

This commit cleans up the mwm implementation so that the monitor
focus is tracked in a structure with a `current` and `next` member,
similar to how it is done in the client implementation.