mwm: Unset monitor focus if the last monitor is detached
When the focused monitor is detached, mwm sets the focus to the next
monitor in a ring list. When the last monitor is detached, mwm will
thus pick the monitor that is about to be detached from the ring list
and focus on it, causing a use-after-free bug.

This commit modifies the `mwm_detach_monitor()` method so that it
checks if the next monitor is the same as the one it is detaching,
and unsets the monitor focus if it is.