monitor: Double-buffer the monitor geometry
The monitor geometry is not double-buffered, causing the geometry
to be changed directly during the event handling, and making it
non-trivial to understand if an event handler is using an old or a
new geometry.

This commit modifies the monitor implementation so that it double
buffers geometries, similar to the client implementation.