monitor: Use the term "crtc" instead of "id"
Monitors are identified using the xid of their crtc as reported by
XRandR. This identifier is called "id", which clashes with monitor
ids that are planned to replace pointers to monitor structures.

This commit modifies mwm so that it uses the term "crtc" for a
monitor's crtc's xid, instead of "id".