monitor: Pass a `struct geom` to `monitor_new()` instead of four integers
The `monitor_new()` method expects to be passed four integers for the
monitor geometry, even though there is a type `struct geom` for this
purpose.

This commit modifies the `monitor_new()` method and its only call-site
so that the method is passed a `struct geom` instead of four separate
integers.