mwm: Fix call to find unviewed workspace when attaching a new monitor
When attaching a new monitor, mwm looks for a workspace that is not
displayed on any monitor and assigns it to the new monitor. However,
because the call to find a workspace is incorrect, it returns the
workspace displayed on the first monitor, or no workspace at all.

This commit modifies the `mwm_attach_monitor()` method so that it
returns the first workspace that is not displayed on any monitor.