workspace: Shift the correct client in `workspace_shift_client()`
Because the type of the `client` parameter has changed from a pointer
to an integer type, the `workspace_shift_client()` method does not
correctly determine whether the specified client or the client that
is currently focused should be shifted.

This commit modifies the `workspace_shift_client()` method so that it
correctly checks if the `client` parameter contains a valid client id.