monitor: Redraw workspace when it is assigned to a monitor
When assigning a workspace to a monitor, `monitor_set_workspace()`
does not mark the workspace as needing to be redrawn, causing the
clients on the workspace not to become visible when a monitor is
changing workspaces.

This commit modifies the `monitor_set_workspace()` method so that
it marks the workspace that is assigned to the monitor as needing
to be redrawn.