mwm: Allow font on indicator bars to be configured separately
Mwm uses one font in the same size for all text that is rendered.
However, since the indicator bars may be used to render kanji, they
require a larger font than the status bar.

This commit modifies mwm so that separate fonts are used to render
text on the status bar and indicator bars.