mwm: Unify event handling with an event queue
When mwm handles an event that refers to the current state of the
window manager, for example the mouse pointer entering a client,
it determines the client under the pointer at the time the event
is handled. If the state has changed between the time the event
was received and the time it is handled, the event will affect the
wrong client.

This client modifies the mwm main loop so that it converts XEvents
at the time they are received into mwm-specific events, including
resource lookups, and enqueues the events so that they are handled
at a later stage of the main loop.
This design also makes it easier to debug mwm because the contents
of the event loop may be dumped more easily.