monitor: Initialize workspace member of monitor structure
The `workspace` member of monitor structures is not initialized
correctly, causing the `monitor_get_workspace()` method to return
a valid workspace id when it is called before the workspace was
set.

This commit modifies the `monitor_new()` method so that the
`workspace` member is initialized with an invalid workspace id.