client: Initialize workspace member of client structure
The `workspace` member of client structures is not initialized by
`client_new()` causing the `client_get_workspace()` method to
return a valid workspace id when it is called before the client's
workspace was set.

This commit modifies the `client_new()` method so that it sets
the `workspace` member to an invalid workspace id when the client
structure is initialized.