monitor: Don't make indicators visible when resizing them
The `_indicator_update_geometry()` method updates the position of
the indicators when resizing them, regardless of their visibility,
causing invisible indiators to be made visible for a brief moment,
which in turn causes the X Server to send XExposeEvent messages
and mwm to trigger unnecessary redraws.

This commit modifies the indicator implementation so that the
visibility of the indicators is remembered. This allows the
`_indicator_update_geometry()` method to determine how it should
update the indicator position.