monitor: Add method to normalize monitor ids
If a monitor id is negative or larger than the highest monitor id, it is
automatically wrapped to the valid range, making it impossible to pass
invalid ids to monitor methods. However, because multiple ids may refer
to the same monitor, this makes it impossible to compare monitors by
their id.

This commit adds the `monitor_normalize_id()` method which normalizes a
monitor id to the effective range of ids, allowing the resulting ids to
be compared.